refactor: 예약 도메인 중심으로 책임 분
