+++ src/main/java/roomescape/service/ReservationService.java
@@ -1,8 +1,9 @@
 package roomescape.service;
 
 import java.time.LocalDate;
+import java.util.ArrayList;
 import java.util.List;
-import java.util.Objects;
+import java.util.Set;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -11,8 +12,6 @@ import roomescape.domain.ReservationTime;
 import roomescape.domain.Theme;
 import roomescape.global.exception.DuplicateEntityException;
 import roomescape.global.exception.EntityNotFoundException;
-import roomescape.global.exception.ForbiddenException;
-import roomescape.global.exception.InactiveException;
 import roomescape.repository.ReservationRepository;
 import roomescape.repository.ReservationTimeRepository;
 import roomescape.repository.ThemeRepository;
@@ -20,6 +19,7 @@ import roomescape.web.dto.reservation.ReservationCancelRequest;
 import roomescape.web.dto.reservation.ReservationModifyRequest;
 import roomescape.web.dto.reservation.ReservationRequest;
 import roomescape.web.dto.reservation.ReservationResponse;
+import roomescape.web.dto.theme.ReservationTimeStatusResponse;
 
 @Service
 @Transactional(readOnly = true)
@@ -35,9 +35,10 @@ public class ReservationService {
         Theme theme = findThemeOrThrow(request.themeId());
         ReservationTime time = findTimeOrThrow(request.timeId());
 
+        theme.validateInactive();
+        time.validateInactiveTime();
+
         validateDuplicateReservation(request.date(), time, theme);
-        validateInactiveTheme(theme);
-        validateInactiveTime(time);
 
         Reservation reservation = Reservation.of(request.name(), request.date(), theme, time);
         return ReservationResponse.from(reservationRepository.save(reservation));
@@ -47,26 +48,17 @@ public class ReservationService {
     public void cancel(Long id, ReservationCancelRequest request) {
         Reservation reservation = findReservationOrThrow(id);
 
-        if (!request.name().equals(reservation.getName())) {
-            throw new ForbiddenException("예약자 명이 일치하지 않습니다.");
-        }
-
+        reservation.validateOwner(request.name());
         reservation.cancel();
         reservationRepository.update(reservation);
     }
 
     public List<ReservationResponse> getAllReservationsByPaging(int page, int size) {
-        return reservationRepository.findAllByPaging(page, size)
-                .stream()
-                .map(ReservationResponse::from)
-                .toList();
+        return reservationRepository.findAllByPaging(page, size).stream().map(ReservationResponse::from).toList();
     }
 
     public List<ReservationResponse> getReservationsByUser(String name) {
-        return reservationRepository.findAllByUserName(name)
-                .stream()
-                .map(ReservationResponse::from)
-                .toList();
+        return reservationRepository.findAllByUserName(name).stream().map(ReservationResponse::from).toList();
     }
 
     @Transactional
@@ -76,19 +68,35 @@ public class ReservationService {
         LocalDate date = request.date();
         Theme theme = reservation.getTheme();
 
-        validateUserName(request.name(), reservation.getName());
+        theme.validateInactive();
+        time.validateInactiveTime();
+
+        reservation.validateOwner(request.name());
         validateDuplicateReservation(date, time, theme);
-        validateInactiveTheme(theme);
-        validateInactiveTime(time);
 
         reservation.update(date, time);
 
         reservationRepository.update(reservation);
     }
 
+    public List<ReservationTimeStatusResponse> getReservationStatusByTheme(Long themeId, LocalDate date) {
+        if (themeRepository.findById(themeId).isEmpty()) {
+            throw new EntityNotFoundException("존재하지 않는 테마 정보입니다.");
+        }
+
+        Set<Long> reservedTimeIds = reservationRepository.findUnavailableTimeIdsByThemeIdAndDate(themeId, date);
+
+        List<ReservationTimeStatusResponse> responses = new ArrayList<>();
+        reservationTimeRepository.findTimeSlotsForReservationStatus().forEach(time -> {
+            boolean reservable = isReservable(time, date, reservedTimeIds);
+            responses.add(ReservationTimeStatusResponse.of(time, reservable));
+        });
+
+        return responses;
+    }
+
     private Theme findThemeOrThrow(Long themeId) {
-        return themeRepository.findById(themeId)
-                .orElseThrow(() -> new EntityNotFoundException("존재하지 않는 테마 정보입니다."));
+        return themeRepository.findById(themeId).orElseThrow(() -> new EntityNotFoundException("존재하지 않는 테마 정보입니다."));
     }
 
     private void validateDuplicateReservation(LocalDate date, ReservationTime time, Theme theme) {
@@ -97,31 +105,16 @@ public class ReservationService {
         }
     }
 
-    private void validateUserName(String name, String reservedName) {
-        if (!Objects.equals(name, reservedName)) {
-            throw new ForbiddenException("예약자 명이 일치하지 않습니다.");
-        }
+    private boolean isReservable(ReservationTime time, LocalDate date, Set<Long> reservedTimeIds) {
+        return time.isAvailableAt(date) && !reservedTimeIds.contains(time.getId()) && time.isActive();
     }
 
     private Reservation findReservationOrThrow(Long id) {
-        return reservationRepository.findById(id)
-                .orElseThrow(() -> new EntityNotFoundException("존재하지 않는 예약입니다."));
+        return reservationRepository.findById(id).orElseThrow(() -> new EntityNotFoundException("존재하지 않는 예약입니다."));
     }
 
     private ReservationTime findTimeOrThrow(Long timeId) {
         return reservationTimeRepository.findById(timeId)
                 .orElseThrow(() -> new EntityNotFoundException("존재하지 않는 시간 정보입니다."));
-    }
-
-    private void validateInactiveTheme(Theme theme) {
-        if (!theme.isActive()) {
-            throw new InactiveException("비활성화 된 테마는 예약할 수 없습니다.");
-        }
-    }
-
-    private void validateInactiveTime(ReservationTime time) {
-        if (!time.isActive()) {
-            throw new InactiveException("비활성화 된 시간대는 예약할 수 없습니다.");
-        }
     }
 }
