+++ src/main/java/roomescape/service/ThemeService.java
@@ -1,22 +1,16 @@
 package roomescape.service;
 
 import java.time.LocalDate;
-import java.util.ArrayList;
 import java.util.List;
-import java.util.Set;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import roomescape.domain.ReservationTime;
 import roomescape.domain.Theme;
 import roomescape.global.exception.DuplicateEntityException;
 import roomescape.global.exception.EntityNotFoundException;
-import roomescape.repository.ReservationRepository;
-import roomescape.repository.ReservationTimeRepository;
 import roomescape.repository.ThemeRepository;
 import roomescape.web.dto.theme.ThemeRequest;
 import roomescape.web.dto.theme.ThemeResponse;
-import roomescape.web.dto.theme.ThemeTimesResponse;
 
 @Service
 @Transactional(readOnly = true)
@@ -24,8 +18,6 @@ import roomescape.web.dto.theme.ThemeTimesResponse;
 public class ThemeService {
 
     private final ThemeRepository themeRepository;
-    private final ReservationRepository reservationRepository;
-    private final ReservationTimeRepository reservationTimeRepository;
 
     @Transactional
     public ThemeResponse register(ThemeRequest request) {
@@ -43,23 +35,6 @@ public class ThemeService {
         themeRepository.update(theme);
     }
 
-    public List<ThemeTimesResponse> getThemeReservationStatus(Long id, LocalDate date) {
-        if (themeRepository.findById(id).isEmpty()) {
-            throw new EntityNotFoundException("존재하지 않는 테마 정보입니다.");
-        }
-
-        Set<Long> reservedTimeIds = reservationRepository.findUnavailableTimeIdsByThemeIdAndDate(id, date);
-
-        List<ThemeTimesResponse> responses = new ArrayList<>();
-        reservationTimeRepository.findTimeSlotsForReservationStatus().forEach(
-                time -> {
-                    boolean reservable = isReservable(time, date, reservedTimeIds);
-                    responses.add(ThemeTimesResponse.of(time, reservable));
-                });
-
-        return responses;
-    }
-
     public List<ThemeResponse> getAllActiveThemesByPaging(int page, int size) {
         return themeRepository.findAllActiveThemesByPaging(page, size)
                 .stream()
@@ -72,10 +47,6 @@ public class ThemeService {
                 .stream()
                 .map(ThemeResponse::from)
                 .toList();
-    }
-
-    private boolean isReservable(ReservationTime time, LocalDate date, Set<Long> reservedTimeIds) {
-        return time.isAvailableAt(date) && !reservedTimeIds.contains(time.getId()) && time.isActive();
     }
 
     private void validateDuplicateName(String name) {
