+++ src/main/java/roomescape/web/controller/user/ThemeController.java
@@ -7,13 +7,11 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.service.ThemeService;
 import roomescape.web.dto.theme.ThemeResponses;
-import roomescape.web.dto.theme.ThemeTimeResponses;
 
 @RestController
 @RequestMapping("/api/themes")
@@ -22,17 +20,6 @@ import roomescape.web.dto.theme.ThemeTimeResponses;
 public class ThemeController {
 
     private final ThemeService themeService;
-
-    @GetMapping("/{id}/times")
-    public ResponseEntity<ThemeTimeResponses> getThemeReservationStatus(
-            @PathVariable
-            @Positive(message = "테마 조회 식별자는 양수여야 합니다.") Long id,
-            @RequestParam LocalDate date
-    ) {
-        ThemeTimeResponses response = new ThemeTimeResponses(themeService.getThemeReservationStatus(id, date));
-
-        return ResponseEntity.ok().body(response);
-    }
 
     @GetMapping
     public ResponseEntity<ThemeResponses> getAllThemesByPaging(
