+++ src/main/java/roomescape/web/dto/theme/ReservationTimeStatusResponses.java
@@ -1,0 +1,8 @@
+package roomescape.web.dto.theme;
+
+import java.util.List;
+
+public record ReservationTimeStatusResponses(
+        List<ReservationTimeStatusResponse> responses
+) {
+}
