--- src/main/java/roomescape/web/dto/theme/ThemeTimeResponses.java
@@ -1,8 +1,0 @@
-package roomescape.web.dto.theme;
-
-import java.util.List;
-
-public record ThemeTimeResponses(
-        List<ThemeTimesResponse> responses
-) {
-}
