+++ src/test/java/roomescape/controller/user/ThemeControllerTest.java
@@ -3,18 +3,14 @@ package roomescape.controller.user;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.hamcrest.Matchers.containsString;
 import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.anyLong;
 import static org.mockito.Mockito.when;
 
 import io.restassured.common.mapper.TypeRef;
 import io.restassured.module.mockmvc.RestAssuredMockMvc;
 import java.time.LocalDate;
-import java.time.LocalTime;
 import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.ValueSource;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.http.HttpStatus;
 import org.springframework.test.context.bean.override.mockito.MockitoBean;
@@ -24,8 +20,6 @@ import roomescape.service.ThemeService;
 import roomescape.web.controller.user.ThemeController;
 import roomescape.web.dto.theme.ThemeResponse;
 import roomescape.web.dto.theme.ThemeResponses;
-import roomescape.web.dto.theme.ThemeTimeResponses;
-import roomescape.web.dto.theme.ThemeTimesResponse;
 
 @WebMvcTest(ThemeController.class)
 class ThemeControllerTest extends BaseControllerUnitTest {
@@ -36,51 +30,6 @@ class ThemeControllerTest extends BaseControllerUnitTest {
     @BeforeEach
     void setUp(WebApplicationContext webApplicationContext) {
         mockMvcSetting(webApplicationContext);
-    }
-
-    @Test
-    void 테마_시간대_조회_요청에_성공하면_200_OK와_정상_응답이_반환된다() {
-        // given
-        ThemeTimeResponses expected = new ThemeTimeResponses(
-                List.of(
-                        new ThemeTimesResponse(1L, LocalTime.of(10, 0), true),
-                        new ThemeTimesResponse(2L, LocalTime.of(11, 0), true)
-                )
-        );
-        when(themeService.getThemeReservationStatus(anyLong(), any(LocalDate.class))).thenReturn(expected.responses());
-
-        // when & then
-        ThemeTimeResponses response = RestAssuredMockMvc.given().spec(defaultSpec()).log().all()
-                .queryParam("date", "2026-05-06")
-                .when().get("/api/themes/1/times")
-                .then().log().all()
-                .status(HttpStatus.OK)
-                .extract().as(new TypeRef<>() {
-                });
-
-        assertThat(response).isEqualTo(expected);
-    }
-
-    @ParameterizedTest
-    @ValueSource(longs = {-1, 0})
-    void 테마_조회_요청_시_테마_식별자가_양수가_아니라면_400_BAD_REQUEST를_응답한다(Long invalidThemeId) {
-        // when & then
-        RestAssuredMockMvc.given().spec(defaultSpec()).log().all()
-                .queryParam("date", "2026-05-06")
-                .when().get("/api/themes/" + invalidThemeId + "/times")
-                .then().log().all()
-                .status(HttpStatus.BAD_REQUEST)
-                .body(containsString("테마 조회 식별자는 양수여야 합니다."));
-    }
-
-    @Test
-    void 테마_시간대_조회_요청_시_날짜가_없으면_400_BAD_REQUEST를_응답한다() {
-        // when & then
-        RestAssuredMockMvc.given().spec(defaultSpec()).log().all()
-                .when().get("/api/themes/1/times")
-                .then().log().all()
-                .status(HttpStatus.BAD_REQUEST)
-                .body(containsString("date 파라미터가 누락 되었습니다."));
     }
 
     @Test
